fix(week-calendar): resolve mypy errors from int->float hours change
CI's mypy caught type mismatches the float widening introduced:
- capacity_service._ticket_to_dict_for_dev params int -> float
- admin total_hours / workitems week_allocated accumulators init 0.0
- pulse forecast/actuals path (_scoped_hours, proj_fc/proj_act, _row) made
  float end-to-end; _row now rounds to 2dp instead of int()-truncating
  (consistent with the no-truncation hours contract)
- time_blocks reassign: narrow Optional before assigning to target_item

No schema/type drift; affected backend tests green.

+++ backend/services/capacity_service.py
@@ -79,10 +79,10 @@ def _ticket_belongs_this_week(item, week_start: datetime, week_end: datetime) ->
 
 def _ticket_to_dict_for_dev(
     item,
-    counted: int,
+    counted: float,
     basis: str,
-    logged_this_week: int,
-    total_logged: int,
+    logged_this_week: float,
+    total_logged: float,
 ) -> dict:
     estimated = item.estimated_hours or 0
     return {
